pyproject.toml
[tool.poetry]
name = "retailer_sales_forecasting_system"
version = "0.1.0"
description = ""
authors = ["Cesar Martinez <cesarcmartinez@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
pandas = "^2.3.3"
numpy = "^2.4.0"
matplotlib = "^3.10.8"
scikit-learn = "^1.8.0"
sqlalchemy = "^2.0.45"
psycopg2-binary = "^2.9.11"
python-dotenv = "^1.2.1"
ipykernel = "^7.1.0"
jupyter = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
